backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
python-docx
chromadb>=0.6.0
python-jose[cryptography]
passlib[bcrypt]
celery
redis
httpx
python-multipart
numpy>=2.0
pydantic[email]
email-validator

# NEW V2 Dependencies
spacy==3.8.13
google-generativeai==0.3.1